// components/schemes_configuration_ui/src/Shared/Types/report.ts
export enum ReportLinkType {
	EmergencyEvents = 9,
	EmergencyEventsCriterions = 10,
	ActivePowerConsumption = 1,
	InfoAboutNSI = 11,
	ReportForm = 8,
	BalancesDevicesStatus = 12,
	OperInformDay = 21,
	OperInformMonth = 22,
	OperInformTotal = 23,
	UGENaturalGas = 24,
	UGEConstantLog = 25,
	UGECorrectionLog = 26,
	UGEInstrumentation = 27,
	UGELogBook = 28,
	BalancesRetrofittingAccountingNodes = 29,
	BalancesStatisticalProcessing = 30,
}