// components/schemes_configuration_ui/src/Containers/NSIInformationPanel/types.ts
export const enum TabName {
	USER_SETTINGS = 'nsi-user-settings',
	MEASURING_INSTRUMENTS_JOURNAL = 'nsi-measuring-instruments-journal',
	DETAILED_INFORMATION = 'nsi-detailed-information',
	CONFIGURATION_HISTORY = 'nsi-configuration-history',
}

export const enum ControlName {
	EDIT_VALUES = 'nsi-edit-values-control',
	SAVE_VALUES = 'nsi-save-values-control',
	SETTINGS = 'nsi-settings-control',
	LINK = 'nsi-link-control',
	ADVANCED_FILTER = 'msi-advanced-filter-control',
	REFRESH = 'nsi-refresh-control',
}